Retrieve stopwords

The functions retrieves stopwords from the stopwords package using the ISO-639-1 encoding. For miscellaneous languages data_stopwords_misc are used.

get_stopwords(code)

Arguments

code
A string with the language code of the stopwords.

Value

Character vector with the stopwords or NULL if the language code is unknown.

Examples

get_stopwords("en")

#>    [1] "'ll"             "'tis"            "'twas"           "'ve"            
#>    [5] "10"              "39"              "a"               "a's"            
#>    [9] "able"            "ableabout"       "about"           "above"          
#>   [13] "abroad"          "abst"            "accordance"      "according"      
#>   [17] "accordingly"     "across"          "act"             "actually"       
#>   [21] "ad"              "added"           "adj"             "adopted"        
#>   [25] "ae"              "af"              "affected"        "affecting"      
#>   [29] "affects"         "after"           "afterwards"      "ag"             
#>   [33] "again"           "against"         "ago"             "ah"             
#>   [37] "ahead"           "ai"              "ain't"           "aint"           
#>   [41] "al"              "all"             "allow"           "allows"         
#>   [45] "almost"          "alone"           "along"           "alongside"      
#>   [49] "already"         "also"            "although"        "always"         
#>   [53] "am"              "amid"            "amidst"          "among"          
#>   [57] "amongst"         "amoungst"        "amount"          "an"             
#>   [61] "and"             "announce"        "another"         "any"            
#>   [65] "anybody"         "anyhow"          "anymore"         "anyone"         
#>   [69] "anything"        "anyway"          "anyways"         "anywhere"       
#>   [73] "ao"              "apart"           "apparently"      "appear"         
#>   [77] "appreciate"      "appropriate"     "approximately"   "aq"             
#>   [81] "ar"              "are"             "area"            "areas"          
#>   [85] "aren"            "aren't"          "arent"           "arise"          
#>   [89] "around"          "arpa"            "as"              "aside"          
#>   [93] "ask"             "asked"           "asking"          "asks"           
#>   [97] "associated"      "at"              "au"              "auth"           
#>  [101] "available"       "aw"              "away"            "awfully"        
#>  [105] "az"              "b"               "ba"              "back"           
#>  [109] "backed"          "backing"         "backs"           "backward"       
#>  [113] "backwards"       "bb"              "bd"              "be"             
#>  [117] "became"          "because"         "become"          "becomes"        
#>  [121] "becoming"        "been"            "before"          "beforehand"     
#>  [125] "began"           "begin"           "beginning"       "beginnings"     
#>  [129] "begins"          "behind"          "being"           "beings"         
#>  [133] "believe"         "below"           "beside"          "besides"        
#>  [137] "best"            "better"          "between"         "beyond"         
#>  [141] "bf"              "bg"              "bh"              "bi"             
#>  [145] "big"             "bill"            "billion"         "biol"           
#>  [149] "bj"              "bm"              "bn"              "bo"             
#>  [153] "both"            "bottom"          "br"              "brief"          
#>  [157] "briefly"         "bs"              "bt"              "but"            
#>  [161] "buy"             "bv"              "bw"              "by"             
#>  [165] "bz"              "c"               "c'mon"           "c's"            
#>  [169] "ca"              "call"            "came"            "can"            
#>  [173] "can't"           "cannot"          "cant"            "caption"        
#>  [177] "case"            "cases"           "cause"           "causes"         
#>  [181] "cc"              "cd"              "certain"         "certainly"      
#>  [185] "cf"              "cg"              "ch"              "changes"        
#>  [189] "ci"              "ck"              "cl"              "clear"          
#>  [193] "clearly"         "click"           "cm"              "cmon"           
#>  [197] "cn"              "co"              "co."             "com"            
#>  [201] "come"            "comes"           "computer"        "con"            
#>  [205] "concerning"      "consequently"    "consider"        "considering"    
#>  [209] "contain"         "containing"      "contains"        "copy"           
#>  [213] "corresponding"   "could"           "could've"        "couldn"         
#>  [217] "couldn't"        "couldnt"         "course"          "cr"             
#>  [221] "cry"             "cs"              "cu"              "currently"      
#>  [225] "cv"              "cx"              "cy"              "cz"             
#>  [229] "d"               "dare"            "daren't"         "darent"         
#>  [233] "date"            "de"              "dear"            "definitely"     
#>  [237] "describe"        "described"       "despite"         "detail"         
#>  [241] "did"             "didn"            "didn't"          "didnt"          
#>  [245] "differ"          "different"       "differently"     "directly"       
#>  [249] "dj"              "dk"              "dm"              "do"             
#>  [253] "does"            "doesn"           "doesn't"         "doesnt"         
#>  [257] "doing"           "don"             "don't"           "done"           
#>  [261] "dont"            "doubtful"        "down"            "downed"         
#>  [265] "downing"         "downs"           "downwards"       "due"            
#>  [269] "during"          "dz"              "e"               "each"           
#>  [273] "early"           "ec"              "ed"              "edu"            
#>  [277] "ee"              "effect"          "eg"              "eh"             
#>  [281] "eight"           "eighty"          "either"          "eleven"         
#>  [285] "else"            "elsewhere"       "empty"           "end"            
#>  [289] "ended"           "ending"          "ends"            "enough"         
#>  [293] "entirely"        "er"              "es"              "especially"     
#>  [297] "et"              "et-al"           "etc"             "even"           
#>  [301] "evenly"          "ever"            "evermore"        "every"          
#>  [305] "everybody"       "everyone"        "everything"      "everywhere"     
#>  [309] "ex"              "exactly"         "example"         "except"         
#>  [313] "f"               "face"            "faces"           "fact"           
#>  [317] "facts"           "fairly"          "far"             "farther"        
#>  [321] "felt"            "few"             "fewer"           "ff"             
#>  [325] "fi"              "fifteen"         "fifth"           "fifty"          
#>  [329] "fify"            "fill"            "find"            "finds"          
#>  [333] "fire"            "first"           "five"            "fix"            
#>  [337] "fj"              "fk"              "fm"              "fo"             
#>  [341] "followed"        "following"       "follows"         "for"            
#>  [345] "forever"         "former"          "formerly"        "forth"          
#>  [349] "forty"           "forward"         "found"           "four"           
#>  [353] "fr"              "free"            "from"            "front"          
#>  [357] "full"            "fully"           "further"         "furthered"      
#>  [361] "furthering"      "furthermore"     "furthers"        "fx"             
#>  [365] "g"               "ga"              "gave"            "gb"             
#>  [369] "gd"              "ge"              "general"         "generally"      
#>  [373] "get"             "gets"            "getting"         "gf"             
#>  [377] "gg"              "gh"              "gi"              "give"           
#>  [381] "given"           "gives"           "giving"          "gl"             
#>  [385] "gm"              "gmt"             "gn"              "go"             
#>  [389] "goes"            "going"           "gone"            "good"           
#>  [393] "goods"           "got"             "gotten"          "gov"            
#>  [397] "gp"              "gq"              "gr"              "great"          
#>  [401] "greater"         "greatest"        "greetings"       "group"          
#>  [405] "grouped"         "grouping"        "groups"          "gs"             
#>  [409] "gt"              "gu"              "gw"              "gy"             
#>  [413] "h"               "had"             "hadn't"          "hadnt"          
#>  [417] "half"            "happens"         "hardly"          "has"            
#>  [421] "hasn"            "hasn't"          "hasnt"           "have"           
#>  [425] "haven"           "haven't"         "havent"          "having"         
#>  [429] "he"              "he'd"            "he'll"           "he's"           
#>  [433] "hed"             "hell"            "hello"           "help"           
#>  [437] "hence"           "her"             "here"            "here's"         
#>  [441] "hereafter"       "hereby"          "herein"          "heres"          
#>  [445] "hereupon"        "hers"            "herself"         "herse”"         
#>  [449] "hes"             "hi"              "hid"             "high"           
#>  [453] "higher"          "highest"         "him"             "himself"        
#>  [457] "himse”"          "his"             "hither"          "hk"             
#>  [461] "hm"              "hn"              "home"            "homepage"       
#>  [465] "hopefully"       "how"             "how'd"           "how'll"         
#>  [469] "how's"           "howbeit"         "however"         "hr"             
#>  [473] "ht"              "htm"             "html"            "http"           
#>  [477] "hu"              "hundred"         "i"               "i'd"            
#>  [481] "i'll"            "i'm"             "i've"            "i.e."           
#>  [485] "id"              "ie"              "if"              "ignored"        
#>  [489] "ii"              "il"              "ill"             "im"             
#>  [493] "immediate"       "immediately"     "importance"      "important"      
#>  [497] "in"              "inasmuch"        "inc"             "inc."           
#>  [501] "indeed"          "index"           "indicate"        "indicated"      
#>  [505] "indicates"       "information"     "inner"           "inside"         
#>  [509] "insofar"         "instead"         "int"             "interest"       
#>  [513] "interested"      "interesting"     "interests"       "into"           
#>  [517] "invention"       "inward"          "io"              "iq"             
#>  [521] "ir"              "is"              "isn"             "isn't"          
#>  [525] "isnt"            "it"              "it'd"            "it'll"          
#>  [529] "it's"            "itd"             "itll"            "its"            
#>  [533] "itself"          "itse”"           "ive"             "j"              
#>  [537] "je"              "jm"              "jo"              "join"           
#>  [541] "jp"              "just"            "k"               "ke"             
#>  [545] "keep"            "keeps"           "kept"            "keys"           
#>  [549] "kg"              "kh"              "ki"              "kind"           
#>  [553] "km"              "kn"              "knew"            "know"           
#>  [557] "known"           "knows"           "kp"              "kr"             
#>  [561] "kw"              "ky"              "kz"              "l"              
#>  [565] "la"              "large"           "largely"         "last"           
#>  [569] "lately"          "later"           "latest"          "latter"         
#>  [573] "latterly"        "lb"              "lc"              "least"          
#>  [577] "length"          "less"            "lest"            "let"            
#>  [581] "let's"           "lets"            "li"              "like"           
#>  [585] "liked"           "likely"          "likewise"        "line"           
#>  [589] "little"          "lk"              "ll"              "long"           
#>  [593] "longer"          "longest"         "look"            "looking"        
#>  [597] "looks"           "low"             "lower"           "lr"             
#>  [601] "ls"              "lt"              "ltd"             "lu"             
#>  [605] "lv"              "ly"              "m"               "ma"             
#>  [609] "made"            "mainly"          "make"            "makes"          
#>  [613] "making"          "man"             "many"            "may"            
#>  [617] "maybe"           "mayn't"          "maynt"           "mc"             
#>  [621] "md"              "me"              "mean"            "means"          
#>  [625] "meantime"        "meanwhile"       "member"          "members"        
#>  [629] "men"             "merely"          "mg"              "mh"             
#>  [633] "microsoft"       "might"           "might've"        "mightn't"       
#>  [637] "mightnt"         "mil"             "mill"            "million"        
#>  [641] "mine"            "minus"           "miss"            "mk"             
#>  [645] "ml"              "mm"              "mn"              "mo"             
#>  [649] "more"            "moreover"        "most"            "mostly"         
#>  [653] "move"            "mp"              "mq"              "mr"             
#>  [657] "mrs"             "ms"              "msie"            "mt"             
#>  [661] "mu"              "much"            "mug"             "must"           
#>  [665] "must've"         "mustn't"         "mustnt"          "mv"             
#>  [669] "mw"              "mx"              "my"              "myself"         
#>  [673] "myse”"           "mz"              "n"               "na"             
#>  [677] "name"            "namely"          "nay"             "nc"             
#>  [681] "nd"              "ne"              "near"            "nearly"         
#>  [685] "necessarily"     "necessary"       "need"            "needed"         
#>  [689] "needing"         "needn't"         "neednt"          "needs"          
#>  [693] "neither"         "net"             "netscape"        "never"          
#>  [697] "neverf"          "neverless"       "nevertheless"    "new"            
#>  [701] "newer"           "newest"          "next"            "nf"             
#>  [705] "ng"              "ni"              "nine"            "ninety"         
#>  [709] "nl"              "no"              "no-one"          "nobody"         
#>  [713] "non"             "none"            "nonetheless"     "noone"          
#>  [717] "nor"             "normally"        "nos"             "not"            
#>  [721] "noted"           "nothing"         "notwithstanding" "novel"          
#>  [725] "now"             "nowhere"         "np"              "nr"             
#>  [729] "nu"              "null"            "number"          "numbers"        
#>  [733] "nz"              "o"               "obtain"          "obtained"       
#>  [737] "obviously"       "of"              "off"             "often"          
#>  [741] "oh"              "ok"              "okay"            "old"            
#>  [745] "older"           "oldest"          "om"              "omitted"        
#>  [749] "on"              "once"            "one"             "one's"          
#>  [753] "ones"            "only"            "onto"            "open"           
#>  [757] "opened"          "opening"         "opens"           "opposite"       
#>  [761] "or"              "ord"             "order"           "ordered"        
#>  [765] "ordering"        "orders"          "org"             "other"          
#>  [769] "others"          "otherwise"       "ought"           "oughtn't"       
#>  [773] "oughtnt"         "our"             "ours"            "ourselves"      
#>  [777] "out"             "outside"         "over"            "overall"        
#>  [781] "owing"           "own"             "p"               "pa"             
#>  [785] "page"            "pages"           "part"            "parted"         
#>  [789] "particular"      "particularly"    "parting"         "parts"          
#>  [793] "past"            "pe"              "per"             "perhaps"        
#>  [797] "pf"              "pg"              "ph"              "pk"             
#>  [801] "pl"              "place"           "placed"          "places"         
#>  [805] "please"          "plus"            "pm"              "pmid"           
#>  [809] "pn"              "point"           "pointed"         "pointing"       
#>  [813] "points"          "poorly"          "possible"        "possibly"       
#>  [817] "potentially"     "pp"              "pr"              "predominantly"  
#>  [821] "present"         "presented"       "presenting"      "presents"       
#>  [825] "presumably"      "previously"      "primarily"       "probably"       
#>  [829] "problem"         "problems"        "promptly"        "proud"          
#>  [833] "provided"        "provides"        "pt"              "put"            
#>  [837] "puts"            "pw"              "py"              "q"              
#>  [841] "qa"              "que"             "quickly"         "quite"          
#>  [845] "qv"              "r"               "ran"             "rather"         
#>  [849] "rd"              "re"              "readily"         "really"         
#>  [853] "reasonably"      "recent"          "recently"        "ref"            
#>  [857] "refs"            "regarding"       "regardless"      "regards"        
#>  [861] "related"         "relatively"      "research"        "reserved"       
#>  [865] "respectively"    "resulted"        "resulting"       "results"        
#>  [869] "right"           "ring"            "ro"              "room"           
#>  [873] "rooms"           "round"           "ru"              "run"            
#>  [877] "rw"              "s"               "sa"              "said"           
#>  [881] "same"            "saw"             "say"             "saying"         
#>  [885] "says"            "sb"              "sc"              "sd"             
#>  [889] "se"              "sec"             "second"          "secondly"       
#>  [893] "seconds"         "section"         "see"             "seeing"         
#>  [897] "seem"            "seemed"          "seeming"         "seems"          
#>  [901] "seen"            "sees"            "self"            "selves"         
#>  [905] "sensible"        "sent"            "serious"         "seriously"      
#>  [909] "seven"           "seventy"         "several"         "sg"             
#>  [913] "sh"              "shall"           "shan't"          "shant"          
#>  [917] "she"             "she'd"           "she'll"          "she's"          
#>  [921] "shed"            "shell"           "shes"            "should"         
#>  [925] "should've"       "shouldn"         "shouldn't"       "shouldnt"       
#>  [929] "show"            "showed"          "showing"         "shown"          
#>  [933] "showns"          "shows"           "si"              "side"           
#>  [937] "sides"           "significant"     "significantly"   "similar"        
#>  [941] "similarly"       "since"           "sincere"         "site"           
#>  [945] "six"             "sixty"           "sj"              "sk"             
#>  [949] "sl"              "slightly"        "sm"              "small"          
#>  [953] "smaller"         "smallest"        "sn"              "so"             
#>  [957] "some"            "somebody"        "someday"         "somehow"        
#>  [961] "someone"         "somethan"        "something"       "sometime"       
#>  [965] "sometimes"       "somewhat"        "somewhere"       "soon"           
#>  [969] "sorry"           "specifically"    "specified"       "specify"        
#>  [973] "specifying"      "sr"              "st"              "state"          
#>  [977] "states"          "still"           "stop"            "strongly"       
#>  [981] "su"              "sub"             "substantially"   "successfully"   
#>  [985] "such"            "sufficiently"    "suggest"         "sup"            
#>  [989] "sure"            "sv"              "sy"              "system"         
#>  [993] "sz"              "t"               "t's"             "take"           
#>  [997] "taken"           "taking"          "tc"              "td"             
#> [1001] "tell"            "ten"             "tends"           "test"           
#> [1005] "text"            "tf"              "tg"              "th"             
#> [1009] "than"            "thank"           "thanks"          "thanx"          
#> [1013] "that"            "that'll"         "that's"          "that've"        
#> [1017] "thatll"          "thats"           "thatve"          "the"            
#> [1021] "their"           "theirs"          "them"            "themselves"     
#> [1025] "then"            "thence"          "there"           "there'd"        
#> [1029] "there'll"        "there're"        "there's"         "there've"       
#> [1033] "thereafter"      "thereby"         "thered"          "therefore"      
#> [1037] "therein"         "therell"         "thereof"         "therere"        
#> [1041] "theres"          "thereto"         "thereupon"       "thereve"        
#> [1045] "these"           "they"            "they'd"          "they'll"        
#> [1049] "they're"         "they've"         "theyd"           "theyll"         
#> [1053] "theyre"          "theyve"          "thick"           "thin"           
#> [1057] "thing"           "things"          "think"           "thinks"         
#> [1061] "third"           "thirty"          "this"            "thorough"       
#> [1065] "thoroughly"      "those"           "thou"            "though"         
#> [1069] "thoughh"         "thought"         "thoughts"        "thousand"       
#> [1073] "three"           "throug"          "through"         "throughout"     
#> [1077] "thru"            "thus"            "til"             "till"           
#> [1081] "tip"             "tis"             "tj"              "tk"             
#> [1085] "tm"              "tn"              "to"              "today"          
#> [1089] "together"        "too"             "took"            "top"            
#> [1093] "toward"          "towards"         "tp"              "tr"             
#> [1097] "tried"           "tries"           "trillion"        "truly"          
#> [1101] "try"             "trying"          "ts"              "tt"             
#> [1105] "turn"            "turned"          "turning"         "turns"          
#> [1109] "tv"              "tw"              "twas"            "twelve"         
#> [1113] "twenty"          "twice"           "two"             "tz"             
#> [1117] "u"               "ua"              "ug"              "uk"             
#> [1121] "um"              "un"              "under"           "underneath"     
#> [1125] "undoing"         "unfortunately"   "unless"          "unlike"         
#> [1129] "unlikely"        "until"           "unto"            "up"             
#> [1133] "upon"            "ups"             "upwards"         "us"             
#> [1137] "use"             "used"            "useful"          "usefully"       
#> [1141] "usefulness"      "uses"            "using"           "usually"        
#> [1145] "uucp"            "uy"              "uz"              "v"              
#> [1149] "va"              "value"           "various"         "vc"             
#> [1153] "ve"              "versus"          "very"            "vg"             
#> [1157] "vi"              "via"             "viz"             "vn"             
#> [1161] "vol"             "vols"            "vs"              "vu"             
#> [1165] "w"               "want"            "wanted"          "wanting"        
#> [1169] "wants"           "was"             "wasn"            "wasn't"         
#> [1173] "wasnt"           "way"             "ways"            "we"             
#> [1177] "we'd"            "we'll"           "we're"           "we've"          
#> [1181] "web"             "webpage"         "website"         "wed"            
#> [1185] "welcome"         "well"            "wells"           "went"           
#> [1189] "were"            "weren"           "weren't"         "werent"         
#> [1193] "weve"            "wf"              "what"            "what'd"         
#> [1197] "what'll"         "what's"          "what've"         "whatever"       
#> [1201] "whatll"          "whats"           "whatve"          "when"           
#> [1205] "when'd"          "when'll"         "when's"          "whence"         
#> [1209] "whenever"        "where"           "where'd"         "where'll"       
#> [1213] "where's"         "whereafter"      "whereas"         "whereby"        
#> [1217] "wherein"         "wheres"          "whereupon"       "wherever"       
#> [1221] "whether"         "which"           "whichever"       "while"          
#> [1225] "whilst"          "whim"            "whither"         "who"            
#> [1229] "who'd"           "who'll"          "who's"           "whod"           
#> [1233] "whoever"         "whole"           "wholl"           "whom"           
#> [1237] "whomever"        "whos"            "whose"           "why"            
#> [1241] "why'd"           "why'll"          "why's"           "widely"         
#> [1245] "width"           "will"            "willing"         "wish"           
#> [1249] "with"            "within"          "without"         "won"            
#> [1253] "won't"           "wonder"          "wont"            "words"          
#> [1257] "work"            "worked"          "working"         "works"          
#> [1261] "world"           "would"           "would've"        "wouldn"         
#> [1265] "wouldn't"        "wouldnt"         "ws"              "www"            
#> [1269] "x"               "y"               "ye"              "year"           
#> [1273] "years"           "yes"             "yet"             "you"            
#> [1277] "you'd"           "you'll"          "you're"          "you've"         
#> [1281] "youd"            "youll"           "young"           "younger"        
#> [1285] "youngest"        "your"            "youre"           "yours"          
#> [1289] "yourself"        "yourselves"      "youve"           "yt"             
#> [1293] "yu"              "z"               "za"              "zero"           
#> [1297] "zm"              "zr"