ESCO occupations bundle

The occupations pillar is one of the three pillars of ESCO. It organizes the occupation concepts. It uses hierarchical relationships between them, metadata as well as mappings to the International Standard Classification of Occupations ISCO in order to structure the occupations. The descriptions of each concept is provided only in English.

* The ESCO version used is ESCO v1 1.0.3 retrieved at 11/12/2019.

occupations_bundle

Format

A data.table with 5 variables, which are:

conceptUri

Uniform Resource Identifier of occupations.

iscoGroup

Four-level classification of occupation groups, see ISCO.

preferredLabel

Preffered name of ESCO occupation concepts.

altLabels

Alternative labels of ESCO occupation concepts.

description

Description of ESCO occupation concepts.

Source

European Skills/Competences, Qualifications and Occupations ESCO.